requirements.txt
discord.py==2.3.2
pymongo==4.6.2
python-dotenv==1.0.0
numpy==1.24.3